Master meter math (#116)
* Can't add duplicate utility entries

* Can't add duplicate gas entries

* Created add building confirmation

* Fixed some bugs

* Master Meter Averaging works with add and update

* Master Meter averaging now works with  add/update building

* master meter averaging now works on csv upload

* Attempt to fix a bug

There is a bug where on the initial boot of the app if the user needs to log into the database the tabs for add and update pages never load until the app is restarted.

* Attempt 2

* Bug Fixed

* Fixed electricity display bug

When a building that is on the master meter is selected in the update page, the electricity cost and usage is no longer cleared. This means the master meter averaging value will now get displayed and won't update to a null value

* Logs added for Master Meter and some light bug fixes